restricte access of php variables to admim
Change:
I am requesting that the "about" page permissions be changed from anyone "-1" to instead a system administrator "SET_UP_PERMISSION".

Reason:
The about displays all of the php env information and is only useful to a system adminstrator. Providing this information to any user exposes data that could be used to exploit the server.